+++ crates/threshold-signature-server/src/attestation/api.rs
@@ -14,13 +14,13 @@
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 use crate::{
-    attestation::errors::AttestationErr,
+    attestation::errors::{AttestationErr, QuoteMeasurementErr},
     chain_api::{entropy, get_api, get_rpc, EntropyConfig},
     helpers::{
         launch::LATEST_BLOCK_NUMBER_ATTEST,
         substrate::{query_chain, submit_transaction},
     },
-    AppState,
+    AppState, SubxtAccountId32,
 };
 use axum::{
     body::Bytes,
@@ -29,10 +29,14 @@ use axum::{
 };
 use entropy_client::user::request_attestation;
 use entropy_kvdb::kv_manager::KvManager;
-use entropy_shared::{OcwMessageAttestationRequest, QuoteContext};
+use entropy_shared::{
+    attestation::{QuoteContext, QuoteInputData, VerifyQuoteError},
+    OcwMessageAttestationRequest,
+};
 use parity_scale_codec::Decode;
 use serde::Deserialize;
-use subxt::tx::PairSigner;
+use subxt::{backend::legacy::LegacyRpcMethods, OnlineClient};
+use tdx_quote::Quote;
 use x25519_dalek::StaticSecret;
 
 /// HTTP POST endpoint to initiate a TDX attestation.
@@ -73,7 +77,9 @@ pub async fn attest(
     // TODO (#1181): since this endpoint is currently only used in tests we don't know what the context should be
     let context = QuoteContext::Validate;
 
-    let quote = create_quote(nonce, &app_state.signer(), &app_state.x25519_secret, context).await?;
+    let quote =
+        create_quote(nonce, app_state.subxt_account_id(), &app_state.x25519_secret, context)
+            .await?;
 
     // Submit the quote
     let attest_tx = entropy::tx().attestation().attest(quote.clone());
@@ -99,7 +105,9 @@ pub async fn get_attest(
 
     let context = context_querystring.as_quote_context()?;
 
-    let quote = create_quote(nonce, &app_state.signer(), &app_state.x25519_secret, context).await?;
+    let quote =
+        create_quote(nonce, app_state.subxt_account_id(), &app_state.x25519_secret, context)
+            .await?;
 
     Ok((StatusCode::OK, quote))
 }
@@ -108,28 +116,23 @@ pub async fn get_attest(
 #[cfg(not(feature = "production"))]
 pub async fn create_quote(
     nonce: [u8; 32],
-    signer: &PairSigner<EntropyConfig, sp_core::sr25519::Pair>,
+    tss_account: SubxtAccountId32,
     x25519_secret: &StaticSecret,
     context: QuoteContext,
 ) -> Result<Vec<u8>, AttestationErr> {
     use rand::{rngs::StdRng, SeedableRng};
     use rand_core::OsRng;
-    use sp_core::Pair;
 
     // In the real thing this is the key used in the quoting enclave
     let signing_key = tdx_quote::SigningKey::random(&mut OsRng);
 
     let public_key = x25519_dalek::PublicKey::from(x25519_secret);
 
-    let input_data = entropy_shared::QuoteInputData::new(
-        signer.signer().public(),
-        *public_key.as_bytes(),
-        nonce,
-        context,
-    );
+    let input_data =
+        QuoteInputData::new(tss_account.clone(), *public_key.as_bytes(), nonce, context);
 
     // This is generated deterministically from TSS account id
-    let mut pck_seeder = StdRng::from_seed(signer.signer().public().0);
+    let mut pck_seeder = StdRng::from_seed(tss_account.0);
     let pck = tdx_quote::SigningKey::random(&mut pck_seeder);
 
     let pck_encoded = tdx_quote::encode_verifying_key(pck.verifying_key())?.to_vec();
@@ -172,18 +175,13 @@ pub async fn validate_new_attestation(
 #[cfg(feature = "production")]
 pub async fn create_quote(
     nonce: [u8; 32],
-    signer: &PairSigner<EntropyConfig, sp_core::sr25519::Pair>,
+    tss_account: SubxtAccountId32,
     x25519_secret: &StaticSecret,
     context: QuoteContext,
 ) -> Result<Vec<u8>, AttestationErr> {
     let public_key = x25519_dalek::PublicKey::from(x25519_secret);
 
-    let input_data = entropy_shared::QuoteInputData::new(
-        signer.signer().public(),
-        *public_key.as_bytes(),
-        nonce,
-        context,
-    );
+    let input_data = QuoteInputData::new(tss_account, *public_key.as_bytes(), nonce, context);
 
     Ok(configfs_tsm::create_quote(input_data.0)
         .map_err(|e| AttestationErr::QuoteGeneration(format!("{:?}", e)))?)
@@ -208,4 +206,26 @@ impl QuoteContextQuery {
             _ => Err(AttestationErr::UnknownContext),
         }
     }
+}
+
+/// Check build-time measurement matches a current-supported release of entropy-tss
+/// This differs slightly from the attestation pallet implementation because here we don't have direct
+/// access to the parameters pallet - we need to make a query
+pub async fn check_quote_measurement(
+    api: &OnlineClient<EntropyConfig>,
+    rpc: &LegacyRpcMethods<EntropyConfig>,
+    quote: &Quote,
+) -> Result<(), QuoteMeasurementErr> {
+    let mrtd_value = quote.mrtd().to_vec();
+    let query = entropy::storage().parameters().accepted_mrtd_values();
+    let accepted_mrtd_values: Vec<_> = query_chain(api, rpc, query, None)
+        .await?
+        .ok_or(QuoteMeasurementErr::NoMeasurementValues)?
+        .into_iter()
+        .map(|v| v.0)
+        .collect();
+    if !accepted_mrtd_values.contains(&mrtd_value) {
+        return Err(VerifyQuoteError::BadMrtdValue.into());
+    };
+    Ok(())
 }